LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/remove-from-document-expected.txt

CONSOLE MESSAGE: line 30: TypeError: null is not an object (evaluating 'sourceIFrame.contentDocument.querySelector("meta").remove')

FAIL A meta must refresh the original document even if it was removed. TypeError: null is not an object (evaluating 'sourceIFrame.contentDocument.querySelector("meta").remove')
